+++ consensus/tests/test_helper.go
@@ -1,4 +1,4 @@
-package consensus
+package tests
 
 import (
 	"bytes"
